Adicionado suporte a sweet alert e data tables

+++ avaliacao-client/src/app/component/abstract/abstract-list-component.ts
@@ -1,7 +1,10 @@
-import {Router} from '@angular/router';
-import {BaseEntity} from '../../model/base-entity';
-import {OnInit} from '@angular/core';
-import {AbstractService} from '../../service/abstract.service';
+import { Router } from '@angular/router';
+import { BaseEntity } from '../../model/base-entity';
+import { OnInit } from '@angular/core';
+import { AbstractService } from '../../service/abstract.service';
+import swal from 'sweetalert2';
+
+declare let $: any;
 
 export abstract class AbstractListComponent<T extends BaseEntity> implements OnInit {
 
@@ -13,6 +16,33 @@ export abstract class AbstractListComponent<T extends BaseEntity> implements OnI
   abstract getService(): AbstractService<T>;
 
   ngOnInit(): void {
+    $(document).ready(() => {
+      $('.table').DataTable({
+        "ordering": false,
+        "language": {
+          "sEmptyTable": "Nenhum registro encontrado",
+          "sProcessing": "A processar...",
+          "sLengthMenu": "Mostrar _MENU_ registos",
+          "sZeroRecords": "Não foram encontrados resultados",
+          "sInfo": "Mostrando de _START_ até _END_ de _TOTAL_ registos",
+          "sInfoEmpty": "Mostrando de 0 até 0 de 0 registos",
+          "sInfoFiltered": "(filtrado de _MAX_ registos no total)",
+          "sInfoPostFix": "",
+          "sSearch": "Procurar:",
+          "sUrl": "",
+          "oPaginate": {
+            "sFirst": "Primeiro",
+            "sPrevious": "Anterior",
+            "sNext": "Seguinte",
+            "sLast": "Último"
+          },
+          "oAria": {
+            "sSortAscending": ": Ordenar colunas de forma ascendente",
+            "sSortDescending": ": Ordenar colunas de forma descendente"
+          }
+        }
+      });
+    });
     this.getService().findAll().subscribe(data => this.list = data as T[]);
   }
 
@@ -29,8 +59,20 @@ export abstract class AbstractListComponent<T extends BaseEntity> implements OnI
   }
 
   delete(entity: T): void {
-    this.getService().delete(entity.id).subscribe(data => {
-      this.list = this.list.filter(item => item.id !== entity.id);
+    swal({
+      title: 'Deseja excluir esse registro?',
+      text: 'isso não pode ser desfeito',
+      confirmButtonText: 'Sim',
+      showCancelButton: true,
+      cancelButtonText: 'Não',
+      focusCancel: true,
+      type: 'question'
+    }).then(choose => {
+      if (choose.value) {
+        this.getService().delete(entity.id).subscribe(data => {
+          this.list = this.list.filter(item => item.id !== entity.id);
+        });
+      }
     });
   }
